backend/requirements.txt
# Backend (i.e. Python) Dependencies:
# This file lists the necessary Python libraries for FastAPI and MongoDB integration.

fastapi # Framework for building APIs
uvicorn # Server to run FastAPI
motor # Async MongoDB driver for FastAPI
pydantic # For data validation and type checking
python-dotenv # To load environment variables from .env file

# for engage scraping
requests
beautifulsoup4